Converted frontend to webpack.
