+++ .eslintrc.json
@@ -1,0 +1,9 @@
+{
+  "env": {
+    "node": true,
+    "es6": true
+  },
+  "parserOptions": {
+    "ecmaVersion": 2018
+  }
+}
