+++ LICENSE.txt
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2017 Kimberli Zhong
+Copyright (c) 2019 Computational Perception & Cognition Group (CSAIL, MIT)
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
