+++ package.json
@@ -5,7 +5,12 @@
   "private": true,
   "scripts": {
     "start": "node ./bin/www",
-    "test": "jest"
+    "build": "webpack --config=webpack.prod.js",
+    "watch": "webpack --config=webpack.dev.js --watch",
+    "dev": "concurrently \"npm run start\" \"npm run watch\"",
+    "prod": "npm run build && npm start",
+    "test": "jest",
+    "lint": "eslint ui-js-src"
   },
   "repository": {
     "type": "git",
@@ -19,15 +24,35 @@
   "homepage": "https://github.com/cfosco/video-mem-annot#readme",
   "dependencies": {
     "cookie-parser": "^1.4.4",
+    "d3-axis": "^1.0.12",
+    "d3-format": "^1.3.2",
+    "d3-scale": "^3.0.0",
+    "d3-selection": "^1.4.0",
+    "d3-shape": "^1.3.4",
     "debug": "^4.1.1",
     "express": "^4.16.4",
     "express-useragent": "^1.0.12",
     "forever": "^0.15.3",
     "morgan": "^1.9.1",
-    "promise-mysql": "^3.3.1"
+    "objectFitPolyfill": "^2.1.1",
+    "promise-mysql": "^3.3.1",
+    "url-search-params-polyfill": "^5.0.0",
+    "uuid": "^3.3.2"
   },
   "devDependencies": {
+    "@babel/core": "^7.3.4",
+    "@babel/preset-env": "^7.3.4",
+    "babel-loader": "^8.0.5",
+    "concurrently": "^4.1.0",
+    "eslint": "^5.15.3",
+    "eslint-plugin-compat": "^3.0.1",
     "jest": "^24.1.0",
-    "supertest": "^3.4.2"
-  }
+    "supertest": "^3.4.2",
+    "webpack": "^4.29.6",
+    "webpack-cli": "^3.3.0",
+    "webpack-merge": "^4.2.1"
+  },
+  "browserslist": [
+    "since 2011"
+  ]
 }
