+++ public/index.html
@@ -12,14 +12,10 @@
 
     <link rel="stylesheet" href="lib/semantic/semantic.min.css">
     <link rel="stylesheet" href="css/mem_annot.css">
-    <script src="lib/url-search-params-polyfill.min.js"></script>
-    <script src="lib/objectFitPolyfill.min.js"></script>
     <script src="lib/jquery/jquery-3.1.1.min.js"></script>
     <script src="lib/jquery/jquery.color-2.1.2.min.js"></script>
     <script src="lib/semantic/semantic.min.js"></script>
-    <script src="lib/d3.v3.min.js"></script>
-    <script src="lib/uuidv4.min.js"></script>
-    <script src="js/custom.js"></script>
+    <script src="js/bundle.js"></script>
 </head>
 
 <body>
