+++ ui-js-src/.eslintrc.json
@@ -1,0 +1,19 @@
+{
+  "env": {
+      "browser": true,
+      "node": false
+  },
+  "parserOptions": {
+    "sourceType": "module"
+  },
+  "plugins": ["compat"],
+  "extends": [
+    "eslint:recommended",
+    "plugin:compat/recommended"
+  ],
+  "settings": {
+    "polyfills": [
+      "urlsearchparams"
+    ]
+  }
+}
