+++ webpack.prod.js
@@ -1,0 +1,6 @@
+const merge = require('webpack-merge');
+const base = require('./webpack.base.js');
+
+module.exports = merge(base, {
+  mode: 'production'
+});
